+++ NEWS.md
@@ -4,6 +4,11 @@
 
 - Reconstruct the 'coors' object, now it is a dataframe rather than list. <2025-04-07, Mon>
 
+## Added
+
+- added `g_layout_poly_sector` function <2025-04-07, Mon>
+- added `spatstat_layout` function <2025-04-07, Mon>
+
 # MetaNet v0.2.2 Notes
 
 ## Added